rename Requestor file
